[DEMO DO NOT MERGE] Visual regression test
The current master branch and the PR are fully generated and compared
pixel by pixel. If there is to much regression the test fails

A small regression is accepted as the time might differ between the
scraping of the pages and changes of only a pixel will not be noticed.