Raise the minimum user password length
A minimum is introduced and enforced both on the server and client side.

Increased the size of the generated passwords with twice the size, 6->12
for users and 16->32 for passwords with more entropy.

Also added function signatures and removed unneeded PHPdoc annotations..

Removed the unneeded null which is set in the configure.