Post the selected identies to the deleteList endpoint
The submit button for this form is rendered outside of this file.

To make sure that when the table places checkboxes we also have a button
to trigger the delete action an extension of the crawler test is needed.
This test makes sure that when the table is loaded in a new template we
also alert to either add an extra button or disable multidelete in case
of entities which cannot be deleted such as in the auditlog.

The inverted test is also added when the button is there that checkboxes
should be rendered.