Revert "Delete contests after multiselect"
This reverts commit d1311a6c3666fdbcdae7772d04385e7c23ce5e4e.

Place checkboxes for multidelete of entities

To make sure that when the table places checkboxes we also have a button
to trigger the delete action an extension of the crawler test is needed.
This test makes sure that when the table is loaded in a new template we
also alert to either add an extra button or disable multidelete in case
of entities which cannot be deleted such as in the auditlog.

Debug statements

Fix closing