Fix CI action: mayhem API tester
We now also trigger on push to show that it works.

To fix the composer packages we now set the environment to development.
This should be fine as we only care for the API output and not the other
workings.

The logs are now changed to be foldable and more information is shown
about the state of the github actions runner. When this works correctly
we could speed this up by removing these.

Also the installation dir is changed to opt to make debugging easier on
which files are placed.