Check if the POST was explicitly confirmed
In later commits we reuse this code with different kinds of posted data.
If the forms have a confirmation variable it must be positive otherwise
we still render the confirmation box. We cannot only check for the value
as this breaks the old behaviour as `->get(parameter)` will yield false
when never sent, so we would never POST for the single delete endpoints.