Fix syntax test running as root
Alternative is to specify to run as domjudge as we do in other tests but
it makes sense to also test this discouraged setup as root in this small
test.

Also check that MySQL max_allowed_packet is larger than output limit.

Even though the testcase in/output could be small, a team could
still generate large output and that could cause MySQL to fail
inserting the data.

Addresses one of the probable causes from #1583.

Also improve the warning message, since it is not triggered by
the testcase size or output limit at all.

Properly make a list of trees

When extending the delete entity to a multiple at some stages we consider the action as a whole and sometimes as individial loops. As deleting is already hard enough I prefer to do the delete in a loop instead of all at once. This makes that we feed that smaller section to the actual function instead of sometimes passing everything and sometimes only looping.

Fix mAPI github action

This is verified locally that it now works again. The config line was
not interpret as multiline, see
https://github.com/ForAllSecure/mapi-action#usage.

Trigger delete event before actually deleting entity.

Remove commented code

I think this was added when make_chroot was copied and refactored as
run_chroot.

Use chroot arch in mirror selection for ubuntu

It's already explained in the comment but:
- Ubuntu has only i386 & amd64 in the archive mirrors and the other
archs are in ports.ubuntu, when selecting where to download packages for
the chroot we should pick the mirror selected for the chroot arch
instead of the host arch as we might install a arm64 chroot in a amd64
machine (why is another question but the script allowes for this).

Don't build chroot based on hostvalues if Distro is set

Implement Jaap his feedback

Test create chroot script for default settings

Test arch flag

Only build in the bats

Test with common archs

Test with different Distributions

Use most common combinations for ARCH*Distro

Test with Distro tags

More tests

Check that arch test can pass

Try to fix the test

Gitlab builds in different dir

Test Releases

Test with Releases

Give jobs more time as this is the tipping point

Quoting is important

Error

Show that Architecture tests pass

Try with other variable

Waste less CI minutes

Use correct test statement

Trigger Distro checks

Add test about chroot dir

We can now also test with Debian

Implicit test the run_chroot and get the internal issue

Restructure to run make_chroot less often