Start testing forcedownload+forceyes
Fedora does not have dpkg resulting in architecture detection not
working. When running the script without arguments on fedora there is an
error which we need to skip first.

We could fix this by downloading dpkg in a similar way as debootstrap and
using that to get the host architecture. Or use uname and convert values
from there to the architecture name used for debian/ubuntu
(x86_64->amd64).