Make table cells fully clickable
See https://github.com/DOMjudge/domjudge/issues/1273 for the full
explaination.

This needs a browserhack for firefox (or the other way around is also
possible) given that most major browsers are now chromium based it makes
sense to use that as default. At the moment using the firefox solution
also for chromium browsers doesn't seem to work. The selector
`td:not(.testcase-results)` is to make sure that we are reasonable
specific to be applied on the tables, it does actually target everything
because the testcase-results class doesn't have anchors in the cells.

All the anchor tags are now stretched to us the full space inside the
TD.

Based on logic in:
https://stackoverflow.com/questions/18488148/how-to-get-div-height-100-inside-td-of-100#comment120502212_18488334