Keep the owner:group of the one downloading the tarball
In the scenario:
./configure --prefix=<something_which_needs_su/sudo>
make domserver <as unpriviliged user which downloaded the tarball>
sudo make-installdomserver <as root>

We would copy the Nelmiodoc files as root, so the user can't run a make
{dist}clean without errors as we can't remove the folder as its owned by
root.

We now also install nelmiodoc together with composer in the base image
for CI:
- together with the needed dependencies
- make sure we get an autoload.php

See
https://github.com/DOMjudge/domjudge/pull/2303#discussion_r1493469404
for the reasoning behind this.

Also fix the maintainer-install option, make would not run if the directory already exists so we would only get to this step if the directory is not there, so removing it would always be a NO-OP.