Reduce the amount of combinations
For the configure tests we check for different versions of ubuntu, as we
multiply here with the number of architectures this would blow up.

A solution for that would be to manually craft those in a copy of this
file but for now that looks like a lot of testing for probably no real
error catching.