Move generation of some CSS to twig
See: https://stackoverflow.com/a/23504796 for some
backgroundinformation. Most of the issues mentioned we already had, we
already generated the CSS inline and now do the same calculations in
another file.

One disadvantage is that we calculate this for every file also if you
don't visit the scoreboard page. We can either fix this by calculating
hte file and later serve it from disk or by providing the ETag or
storing the settings in memory.