Test for both with and without example data
This is to make sure that when we have no entities for something (which
we rarely test as we always upload data) the pages still look decent.

As this script was used before we keep the old behaviour of always
installing the data when there is no explicit database method given.