Run the unit tests
We'll need to play around with the unit test upload features
(https://github.com/EnricoMi/publish-unit-test-result-action#configuration)
what works best.

This is based on what we do for the integration test. Most shared code
is handled by baseinstall.sh.

We can now also install a database with a different name as the unit
tests use that locally. As side-effect we directly test this in CI.

The CI ran on GitLab before and needed extra steps to push the result to
GitHub. As everything is in GitHub Actions now we can remove those extra
steps.

(cherry picked from commit 55301894f6d15108021787ae18b17ff6db5116f1)