Extend the java tests
Extended the default hello world in Java with the explicit function
signature. This was already implicit but it's better to explicit test
for this in case someone uses another JVM.

Added a compiler-error submission which adds it's own implementation of
the `final String` to make sure that we correctly detect this in the
`detect_main`. A similar (correct) example is added to make debugging
easier.

Added a `correct` solution which `detect-main` fails on as both classes
could be used as entrypoint. This would work if the entrypoint is
explicit.

Added an alternative solution for printing, mostly for timing.