Create a simple judgehost_check up to the judgedaemon
This shows a simple button to run the script on the judgehost, when this works
we can do this by default when a judgehost checks in.

Next things:
- Get output from the judgehost
- Store that output
- Diff that output with other judgehosts

In a future PR we can then even disable a judgehost if it is different,
think of a different CPU, different kernel or other software or hardware
differences.